create: expand number of fibers for more sky/fluxcal
We can't do good rejection on sky and fluxcal spectra if there's
only one or two. In order to get a decent number of those and keep
a reasonable number of object fibers, we need a modest increase in
the number of fibers. This uses the new 'fifteen' fiber setting in
the simulator (15 fibers, uniformly spread across the slit).

In the process, noticed that the choice of fibers is really linked
to the objId values, so we can't really have the choice of fibers
be configurable because the objId values aren't configurable; so
removed configurability of the fibers.